refactor: first pass at clippy pedantic warnings for client and lib

+++ rustyhack_client/src/client_consts.rs
@@ -1,8 +1,8 @@
-pub(crate) const CONSOLE_WIDTH: usize = 71;
-pub(crate) const CONSOLE_HEIGHT: usize = 31;
-pub(crate) const VIEWPORT_WIDTH: usize = 51;
-pub(crate) const VIEWPORT_HEIGHT: usize = 19;
-pub(crate) const TARGET_FPS: usize = 10;
+pub(crate) const CONSOLE_WIDTH: u32 = 71;
+pub(crate) const CONSOLE_HEIGHT: u32 = 31;
+pub(crate) const VIEWPORT_WIDTH: u32 = 51;
+pub(crate) const VIEWPORT_HEIGHT: u32 = 19;
+pub(crate) const TARGET_FPS: u32 = 10;
 pub(crate) const LOG_NAME: &str = "rustyhack_client.log";
 pub(crate) const GAME_TITLE: &str = "Rustyhack MMO";
 pub(crate) const VALID_NAME_REGEX: &str = "^[[:alpha:]]+$";
